Retire the cordillera package name
`cordillera` was a library once; its sources became the `telekinesis.http2`
component, but the package name stayed behind. It needed a footnote in two places
to stop it being misread — `build.mill` explained that the sources "retain the
`cordillera` package", and `doc/roadmap/modularity.md` recorded a scan trap where
a package-name grep called a live dependency edge dead because the package did not
match the library. Both notes go with the name.

`package cordillera` becomes `package telekinesis`: the component is
`telekinesis.http2`, two of its eleven files already declared that package, and
none of the nine toplevel names collides with one already in `telekinesis`.

One collision did have to be resolved. Both `cordillera_test.scala` and
`telekinesis_test.scala` declare `object Tests`, in the same test module, so the
HTTP/2 suite becomes `telekinesis.Http2Tests` in `telekinesis.Http2Tests.scala` —
the convention already used for a library's second suite, as in
`jacinta.ParserTests.scala`.

Three consequences of merging two packages into one:

  - `scintillate.Http2Serve.scala` and `obligatory.Grpc.scala` imported both
    packages, so the rename left a duplicate `import telekinesis.*` to collapse.
  - Four files inside the component, and `telekinesis_test.scala`, imported
    `cordillera.*` while themselves being `package telekinesis`; those imports are
    now imports of the enclosing package, and are deleted rather than rewritten.
  - `private[cordillera]` becomes `private[telekinesis]`, which widens those
    members' visibility from one component to the library. There is no narrower
    enclosing package, and the alternative — keeping a package purely to scope
    fourteen access qualifiers — is what this commit removes.

`rep/` is left alone: it is a dated decision log and a compiler-bug reproduction,
and rewriting them would falsify what was true when they were written. The comment
naming a technique "the cordillera recipe" in three files now describes it instead.
