+++ lib/embarcadero/src/containerd/embarcadero.Workload.scala
@@ -35,7 +35,7 @@ package embarcadero
 import anticipation.*
 import aperture.*
 import contingency.*
-import cordillera.*
+import telekinesis.*
 import gossamer.*
 import locomotion.*
 import obligatory.*
