+++ lib/obligatory/src/grpc/obligatory.grpcInternal.scala
@@ -36,7 +36,7 @@ import scala.quoted.*
 
 import anticipation.*
 import contingency.*
-import cordillera.*
+import telekinesis.*
 import distillate.*
 import fulminate.*
 import gigantism.*
