+++ lib/scintillate/src/server/scintillate.Http2Serve.scala
@@ -39,7 +39,6 @@ import anticipation.*
 import coaxial.*
 import contingency.*
 import denominative.*
-import cordillera.*
 import gossamer.*
 import parasite.*
 import prepositional.*
