+++ lib/telekinesis/src/http2/soundness_telekinesis_http2.scala
@@ -30,41 +30,7 @@
 ┃                                                                                                  ┃
 ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
                                                                                                   */
-package cordillera
+package soundness
 
-import java.util.concurrent.locks as jucl
-
-import scala.math
-import scala.caps
-
-import proscenium.*
-
-// A single HTTP/2 flow-control send window (RFC 7540 §6.9): a signed byte
-// budget the peer grants us for sending DATA. `acquire` blocks the sending
-// fiber until at least one byte is available and returns up to `n` (so a large
-// body drains as the window opens); `release` (from an inbound WINDOW_UPDATE)
-// tops it up and wakes waiters. A `ReentrantLock`/`Condition` rather than
-// `synchronized`, so a blocked virtual thread unmounts its carrier.
-class FlowWindow(initial: Int):
-  private val lock: jucl.ReentrantLock = jucl.ReentrantLock()
-  private val replenished: jucl.Condition = lock.newCondition().nn
-
-  // Guarded by `lock`; reached only through this window's own methods.
-  @caps.unsafe.untrackedCaptures
-  private var value: Long = initial.toLong
-
-  def acquire(n: Int): Int =
-    lock.lock()
-    try
-      while value <= 0 do replenished.await()
-      val take: Int = math.min(n.toLong, value).toInt
-      value -= take
-      take
-    finally lock.unlock()
-
-  def release(increment: Int): Unit =
-    lock.lock()
-    try
-      value += increment.toLong
-      replenished.signalAll()
-    finally lock.unlock()
+export telekinesis.{HpackEntry, HpackTable, Http2, Hpack, Huffman, FrameReader,
+    PseudoHeaders}
