+++ lib/telekinesis/src/http2/telekinesis.HttpSession.scala
@@ -42,7 +42,6 @@ import javax.net.ssl as jns
 import anticipation.*
 import coaxial.*
 import contingency.*
-import cordillera.*
 import gigantism.*
 import gossamer.*
 import parasite.*
