+++ lib/telekinesis/src/http2/telekinesis.PseudoHeaders.scala
@@ -30,52 +30,108 @@
 ┃                                                                                                  ┃
 ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
                                                                                                   */
-package cordillera
+package telekinesis
 
-import scala.caps
-import proscenium.compat.*
+import scala.collection.mutable as scm
 
-import anticipation.{Data as Bytes, *}
-import prepositional.*
-import vacuous.*
+import anticipation.*
+import contingency.*
+import distillate.*
+import gossamer.*
 import rudiments.*
+import spectacular.*
+import telekinesis.*
+import urticose.*
+import vacuous.*
+import zephyrine.*
 
-// A minimal growable byte buffer, classified as an exclusive, stateful capability:
-// the separation checker rejects `scala.collection.mutable.ArrayBuilder` mutation
-// outside a `uses`-clause scope, so the HTTP/2 wire encoders append through this
-// instead. `data` copies out, so the internal storage never escapes.
-private[cordillera] class ByteBuf(initial: Int = 32)
-extends caps.ExclusiveCapability, caps.Stateful:
-  // Untracked: reached only through this (exclusive) buffer, and `data` copies out.
-  @caps.unsafe.untrackedCaptures
-  private var storage: scala.Array[Byte] = new scala.Array[Byte](initial.max(8))
-  private var size0: Int = 0
+import Http2.Error.Reason
 
-  // An exclusive view for writes: the untracked field reads as read-only.
-  private inline def target: scala.Array[Byte]^ = storage.asInstanceOf[scala.Array[Byte]^]
+// Translates between telekinesis's transport-agnostic HTTP model and HTTP/2's
+// HPACK header blocks. On the request side the pseudo-headers (`:method`, `:scheme`,
+// `:authority`, `:path`) are emitted first, in order, ahead of the regular headers
+// (RFC 7540 §8.1.2.1); on the response side `:status` becomes an `Http.Status` and
+// the remaining fields become the response headers.
+object PseudoHeaders:
+  // Build the HPACK header list for a request. `scheme`/`authority` come from the
+  // connection context, since an `Http.Request` carries them separately.
+  def request(request: Http.Request, scheme: Text, authority: Text): List[HpackEntry] =
+    val pseudo =
+      List
+        ( HpackEntry(t":method", request.method.show),
+          HpackEntry(t":scheme", scheme),
+          HpackEntry(t":authority", authority),
+          HpackEntry(t":path", request.target) )
 
-  def size: Int = size0
+    val regular = request.textHeaders.map: header => HpackEntry(header.key.lower, header.value)
 
-  private update def ensure(extra: Int): Unit =
-    if size0 + extra > storage.length then
-      var capacity = storage.length*2
-      while size0 + extra > capacity do capacity *= 2
-      val grown = new scala.Array[Byte](capacity)
-      System.arraycopy(storage, 0, grown, 0, size0)
-      // The cast erases the fresh array's capture: it is confined to this buffer.
-      storage = grown.asInstanceOf[scala.Array[Byte]]
+    List.of(pseudo.stdlib ++ regular.stdlib)
 
-  update def add(byte: Byte): Unit =
-    ensure(1)
-    target(size0) = byte
-    size0 += 1
+  // Reconstruct an `Http.Response` from a decoded HEADERS block and the body stream.
+  // `:status` selects the `Http.Status`; other fields become response headers.
+  def response(headerBlock: List[HpackEntry], body: Chain[Data])
+  :   Http.Response raises Http2.Error =
 
-  update def addAll(bytes: Bytes): Unit =
-    ensure(bytes.length)
-    System.arraycopy(Array.unsafeJvm(bytes), 0, target, size0, bytes.length)
-    size0 += bytes.length
+    var statusText: Optional[Text] = Unset
+    val headers = scm.ListBuffer[Http.Header]()
 
-  def data: Bytes =
-    val out = Array[Byte](size0)
-    System.arraycopy(storage, 0, out.raw, 0, size0)
-    Array.freeze(out)
+    headerBlock.each: entry =>
+      if entry.name == t":status" then statusText = entry.value
+      else if !entry.name.starts(t":") then headers += Http.Header(entry.name, entry.value)
+
+    val code: Int = statusText.let { text => safely(Integer.parseInt(text.s)).or(0) }.or(0)
+
+    val status: Http.Status =
+      Http.Status.unapply(code).optional.lest(Http2.Error(Reason.Protocol(t"missing :status")))
+
+    status(headers.to(List), Http.Body.Flowing(() => zephyrine.Stream(body.stdlib.iterator)))
+
+  // Reconstruct an `Http.Request` from a decoded request HEADERS block and the
+  // body spring: `:method`/`:path` select the method and target, `:authority`
+  // the host (any port stripped, as for an HTTP/1.1 `Host` header); other
+  // pseudo-headers are dropped and the remaining fields become request headers.
+  // The inverse of `request`, used by the server role.
+  // Plain using-parameter, de-sugared from `raises`: a context-function result
+  // may not hide the capturing request.
+  def requestOf(headerBlock: List[HpackEntry], consume body: Spring[Data]^)
+    ( using Tactic[Http2.Error] )
+  :   Http.Request^ =
+
+    var methodText: Optional[Text] = Unset
+    var pathText: Optional[Text] = Unset
+    var authorityText: Optional[Text] = Unset
+    val headers = scm.ListBuffer[Http.Header]()
+
+    headerBlock.each: entry =>
+      if entry.name == t":method" then methodText = entry.value
+      else if entry.name == t":path" then pathText = entry.value
+      else if entry.name == t":authority" then authorityText = entry.value
+      else if !entry.name.starts(t":") then headers += Http.Header(entry.name, entry.value)
+
+    val authority: Text =
+      authorityText.lest(Http2.Error(Reason.Protocol(t"missing :authority")))
+
+    val host: Host =
+      safely(authority.as[Host]).or:
+        safely(authority.cut(t":").prim.or(authority).as[Host]).or:
+          abort(Http2.Error(Reason.Protocol(t"bad :authority")))
+
+    val target: Text = pathText.lest(Http2.Error(Reason.Protocol(t"missing :path")))
+    val method: Http.Method = methodText.lest(Http2.Error(Reason.Protocol(t"missing :method"))).as
+
+    Http.Request(method, 2.0, host, target, headers.to(List), body)
+
+  // Build the HPACK header list for a response: `:status` first, then the
+  // regular headers lowercased, with connection-specific headers stripped
+  // (RFC 7540 §8.1.2.2). The inverse of `response`, used by the server role.
+  def entries(response: Http.Response^): List[HpackEntry] =
+    val forbidden: List[Text] =
+      List(t"connection", t"keep-alive", t"transfer-encoding", t"upgrade", t"proxy-connection")
+
+    val regular = response.textHeaders.map: header =>
+      HpackEntry(header.key.lower, header.value)
+
+    . filter: entry =>
+        !forbidden.has(entry.name)
+
+    List.of(HpackEntry(t":status", response.status.code.show) :: regular.stdlib)
