+++ lib/telekinesis/src/http2/telekinesis_http2.scala
@@ -41,7 +41,6 @@ import javax.net.ssl as jns
 import anticipation.*
 import coaxial.*
 import contingency.*
-import cordillera.*
 import gigantism.*
 import gossamer.*
 import parasite.*
